backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
chromadb==0.5.15
pydantic==2.9.2
pillow==10.4.0
